moved functions from menu/portfolio.py to insights/portfolio/formatters.py
