+++ src/db/portfolio.py
@@ -1,7 +1,4 @@
-# src/db/portfolio.py
-
 from __future__ import annotations
-
 import json
 import sqlite3
 from typing import Any, Dict, List, Optional, Tuple
