+++ src/insights/portfolio/__init__.py
@@ -1,0 +1,17 @@
+from .formatters import (
+    format_duration,
+    format_languages,
+    format_frameworks,
+    format_activity_line,
+    format_skills_block,
+    format_summary_block,
+)
+
+__all__ = [
+    "format_duration",
+    "format_languages",
+    "format_frameworks",
+    "format_activity_line",
+    "format_skills_block",
+    "format_summary_block",
+]
